Partially fix talk back reader reads lists wrong for Hydration data type in Nutrtion category
This is issue is addressed partially with this change. Although a PreferenceCategory may lack a title, TalkBack continues to recognize its presence and count it as a member of the list. This behavior leads to an index skip when attempting to focus on the first permission type following a header.
To address this issue effectively, I'm reassigning this bug to the Preference Team for further investigation and resolution.

Bug: 305900531
Test: atest
Change-Id: Ib60a4ac98c58eff1bbf31b58eb31d9af22d41717